스마트에디터 이미지등록 구현중..

+++ src/main/java/com/smarteditor/example/board/dto/AttachDTO.java
@@ -17,6 +17,6 @@ public class AttachDTO extends CommonDTO {
 	
 	private String saveImgName;
 	
-	private int imgSize;
+	private long imgSize;
 	
 }
